make Form comp. and set up handler - display text from input
